# india-exporter-search/docs/API_SPEC.yaml
openapi: 3.0.0
info:
  title: India Exporter Buyer Search API
  description: API for searching and managing exporter/buyer company data
  version: 1.0.0
  contact:
    email: support@exportersearch.com
servers:
  - url: https://api.exportersearch.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/companies/{id}:
    get:
      tags: [Companies]
      summary: Get company details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Company'
                  limited:
                    type: boolean
        '404':
          description: Company not found
...
  /api/subscriptions/checkout:
    post:
      tags: [Subscriptions]
      summary: Create checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId:
                  type: string
                  enum: [MONTHLY, YEARLY, CREDITS]
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      sessionId:
                        type: string
                      url:
                        type: string
...
  /api/admin/companies/{id}/verify:
    post:
      tags: [Admin]
      summary: Verify or reject a company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [VERIFIED, REJECTED]
      responses:
        '200':
          description: Company verification updated
        '403':
          description: Forbidden - Admin only
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Company:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        fullAddress:
          type: string
        country:
          type: string
        city:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        importedProducts:
          type: array
          items:
            type: string
        exportedProducts:
          type: array
          items:
            type: string
        importValue:
          type: number
        exportValue:
          type: number
        categories:
          type: array
          items:
            type: string
        hsnCodes:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [PENDING, VERIFIED, REJECTED, FLAGGED]
        dataSource:
          type: string
        createdAt:
          type: string
          format: date-time
...